+++ src/auth/guards/role.decorator.ts
@@ -1,0 +1,7 @@
+import { Reflector } from '@nestjs/core';
+
+interface Permission {
+  context: string;
+}
+
+export const Role = Reflector.createDecorator<Permission>();