+++ src/user/user.module.ts
@@ -1,0 +1,14 @@
+import { Module } from '@nestjs/common';
+import { UserService } from './user.service';
+import { UserController } from './user.controller';
+import { AuthService } from 'src/auth/auth.service';
+import { AuthModule } from 'src/auth/auth.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { User } from './entities/user.entity';
+
+@Module({
+  imports: [AuthModule, TypeOrmModule.forFeature([User])],
+  controllers: [UserController],
+  providers: [UserService, AuthService],
+})
+export class UserModule {}