+++ src/utils/role.enum.ts
@@ -1,0 +1,4 @@
+export enum RoleType {
+  NEW_URL = 'NEW_URL',
+  USER = 'USER',
+}