+++ CHANGELOG.md
@@ -7,9 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.0] - 2024-08-08
+## [0.1.0] - 2024-08-08
 ### Added
 - NestJS project started.
 - Docker File e Docker Compose setup.
 - TypeORM.
 - Postgres connections setup.
+- User module, create, singIn and update route
+
+## [0.2.0] - 2024-08-09
+### Added
+- Shortener module, create, find and findOne route
+
+## [0.3.0] - 2024-08-09
+### Added
+- AuthGuard
+- Shortener routes finOne exceptions, find all by user, update and delete
+- Swagger api documentation
+
+## [0.4.0] - 2024-08-09
+### Added
+- Swagger api documentation descriptions
+- README
+
+### Changed
+- Api url shortener response from string to a object with a url string